Move i18n-related files under i18n folder
